requirements.txt
openai==1.93.0
pandas==2.2.3
pydantic==2.11.7
pydantic-extra-types==2.10.5
shapely==2.0.6
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
zstandard>=0.21.0
numpy>=1.24.0
scikit-image>=0.21.0
opencv-python>=4.0.0